# docs/generated/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Ahlan Commerce API",
    "description": "REST surface for health, products, and import jobs. The GraphQL surface is documented separately in docs/generated/schema.graphql.",
    "license": {
      "name": ""
    },
    "version": "0.1.0"
  },
  "paths": {
    "/api/import-jobs": {
      "get": {
        "tags": [
          "import-jobs"
        ],
        "operationId": "list_import_jobs",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "description": "Filter by job status: queued, running, succeeded, failed. Defaults to failed.",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Import jobs",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ImportJobView"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Unexpected error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "import-jobs"
        ],
        "operationId": "create_import_job",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateImportJobRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Job accepted and queued",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateImportJobResponse"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
    "/api/products": {
      "get": {
        "tags": [
          "products"
        ],
        "operationId": "list_products",
        "responses": {
          "200": {
            "description": "All products",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProductResponse"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Unexpected error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "products"
        ],
        "operationId": "create_product",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProductCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Product created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductResponse"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "409": {
            "description": "Handle already exists",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
    "/api/products/{id}": {
      "patch": {
        "tags": [
          "products"
        ],
        "operationId": "update_product",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Product UUID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProductUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Product updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid id or body",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "Product not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
    "/api/published_products": {
      "get": {
        "tags": [
          "products"
        ],
        "operationId": "get_published_products",
        "responses": {
          "200": {
            "description": "Published products only",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProductResponse"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Unexpected error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
    "/health": {
      "get": {
        "tags": [
          "health"
        ],
        "operationId": "health",
        "responses": {
          "200": {
            "description": "Service is up",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HealthResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "CreateImportJobRequest": {
        "type": "object",
        "properties": {
          "input_path": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
      "CreateImportJobResponse": {
        "type": "object",
        "required": [
          "job"
        ],
        "properties": {
          "job": {
            "$ref": "#/components/schemas/ImportJobView"
          }
        }
      },
      "ErrorBody": {
        "type": "object",
        "required": [
          "code",
          "message",
          "request_id"
        ],
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "request_id": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
      "ErrorEnvelope": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          }
        }
      },
      "HealthResponse": {
        "type": "object",
        "required": [
          "status"
        ],
        "properties": {
          "status": {
            "type": "string"
          }
        }
      },
      "ImportJobView": {
        "type": "object",
        "required": [
          "id",
          "status"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "status": {
            "type": "string"
          }
        }
      },
      "ProductCreateRequest": {
        "type": "object",
        "required": [
          "title",
          "handle",
          "price_cents",
          "inventory_quantity",
          "published"
        ],
        "properties": {
          "description": {
            "type": [
              "string",
              "null"
            ]
          },
          "handle": {
            "type": "string"
          },
          "inventory_quantity": {
            "type": "integer",
            "format": "int64"
          },
          "price_cents": {
            "type": "integer",
            "format": "int64"
          },
          "published": {
            "type": "boolean"
          },
          "title": {
            "type": "string"
          }
        }
      },
      "ProductResponse": {
        "type": "object",
        "required": [
          "id",
          "title",
          "handle",
          "price_cents",
          "inventory_quantity",
          "published",
          "created_at",
          "updated_at"
        ],
        "properties": {
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "description": {
            "type": [
              "string",
              "null"
            ]
          },
          "handle": {
            "type": "string"
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "inventory_quantity": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "price_cents": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "published": {
            "type": "boolean"
          },
          "published_at": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "title": {
            "type": "string"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "ProductUpdateRequest": {
        "type": "object",
        "required": [
          "published"
        ],
        "properties": {
          "description": {
            "type": [
              "string",
              "null"
            ]
          },
          "published": {
            "type": "boolean"
          }
        }
      }
    }
  },
  "tags": [
    {
      "name": "health",
      "description": "Liveness checks"
    },
    {
      "name": "products",
      "description": "Product catalog REST endpoints"
    },
    {
      "name": "import-jobs",
      "description": "Background product-import jobs"
    }
  ]
}